+++ bootique-jdbc-instrumented/src/test/java/io/bootique/jdbc/instrumented/InstrumentedJdbcModuleProviderTest.java
@@ -3,10 +3,17 @@ package io.bootique.jdbc.instrumented;
 import io.bootique.test.junit.BQModuleProviderChecker;
 import org.junit.Test;
 
-public class InstrumentedJdbcModuleProviderIT {
+import static org.junit.Assert.assertEquals;
+
+public class InstrumentedJdbcModuleProviderTest {
 
 	@Test
 	public void testPresentInJar() {
 		BQModuleProviderChecker.testPresentInJar(InstrumentedJdbcModuleProvider.class);
+	}
+
+	@Test
+	public void testConfigPrefix() {
+		assertEquals("jdbc", new InstrumentedJdbcModule().getConfixPrefix());
 	}
 }