+++ RELEASE-NOTES.md
@@ -2,7 +2,8 @@
 
 * #10 Upgrade to Bootique 0.20
 * #11 Basic Derby-based unit test framework
-* #12  JDBC test utils inspired by Cayenne unit tests
+* #12 JDBC test utils inspired by Cayenne unit tests
+* #14 InstrumentedJdbcModule uses incorrect config key
 
 ## 0.11